Automated Text Extraction/Fastapi/requirements.txt
fastapi
uvicorn
boto3
pydantic
mysql-connector-python
requests
python-dotenv